feat: embedding-service UV migrated
