+++ embedding-service/internal/config/__init__.py
@@ -1,0 +1,3 @@
+from .config import Settings
+
+__all__ = ["Settings"]
