+++ embedding-service/internal/models/__init__.py
@@ -1,0 +1,3 @@
+from .model import TextEmbeddingModel
+
+__all__ = ["TextEmbeddingModel"]
