--- embedding-service/models/model.py
@@ -1,15 +1,0 @@
-from transformers import BertTokenizer, BertModel
-import torch
-
-
-class TextEmbeddingModel:
-    def __init__(self):
-        self.tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
-        self.model = BertModel.from_pretrained("bert-base-uncased")
-
-    def get_embeddings(self, text: str):
-        inputs = self.tokenizer(text, return_tensors="pt")
-        outputs = self.model(**inputs)
-
-        embeddings = outputs.last_hidden_state.mean(dim=1)
-        return embeddings.detach().numpy()
