+++ embedding-service/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "embedding-service"
+version = "1.1.0"
+description = "Embedding Service for Galore"
+readme = "README.md"
+requires-python = ">=3.14"
+dependencies = [
+  "accelerate>=1.12.0",
+  "fastapi[all]>=0.127.0",
+  "hf-xet>=1.2.0",
+  "psutil>=7.2.0",
+  "requests>=2.32.5",
+  "torch>=2.9.1",
+  "transformers>=4.57.3",
+]
